requirements.txt
APScheduler==3.10.1
certifi==2026.1.4
charset-normalizer==3.4.4
dataclasses-json==0.5.8
dotenv==0.9.9
Flask==3.1.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
idna==3.11
marshmallow==3.26.2
marshmallow-enum==1.5.1
mypy_extensions==1.1.0
packaging==25.0
phonepe-pg-sdk-python==2.1.7
python-dotenv==1.2.1
pytz==2025.2
requests==2.31.0
setuptools==80.9.0
six==1.17.0
supabase==2.10.0
typing-inspect==0.9.0
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==1.26.12
